Typecasts the keccak256 hash as a uint.

+++ Contract.sol
@@ -16,7 +16,7 @@ contract ZombieFactory {
     } //end createZombie()
 
     function _generateRandomDna(string memory _str) private view returns(uint) {
-        uint rand = keccak256(abi.encodePacked(_str));
+        uint rand = uint(keccak256(abi.encodePacked(_str)));
         return rand % dnaModulus;
     } //end function _generateRandomDna()
 } //end ZombieFactory {}
